move COFFEE algorithm from TSFC to GEM
Two modes share this: 'coffee' and 'spectral'.
